+++ test/test_client.py
@@ -1095,7 +1095,7 @@ class TestClient(IntegrationTest):
 
         # Simulate an authenticate() call on a different socket.
         credentials = auth._build_credentials_tuple(
-            'DEFAULT', 'admin', db_user, db_pwd, {})
+            'DEFAULT', 'admin', db_user, db_pwd, {}, None)
 
         c._cache_credentials('test', credentials, connect=False)